pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["codebaya <beauglory@gmail.com>"]
readme = "README.md"
packages = [{include = "airbnb_clone_backend"}]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"
PyJWT = "^2.6.0"
django-environ = "^0.9.0"
django-cors-headers = "^3.13.0"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
